+++ canonicalwebteam/discourse_docs/__init__.py
@@ -1,0 +1,1 @@
+from canonicalwebteam.discourse_docs.app import DiscourseExtension  # noqa
