--- canonicalwebteam/docs/__init__.py
@@ -1,1 +1,0 @@
-from canonicalwebteam.docs.app import DiscourseExtension  # noqa
